# src/Presentation/Server/wwwroot/api/v1/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "CleanArchitecture API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Register",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseOfInteger"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Login",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseOfString"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/change-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_ChangePassword",
        "requestBody": {
          "x-name": "newPassword",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseOfBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Category/admin": {
      "get": {
        "tags": [
          "Category"
        ],
        "operationId": "Category_GetAdminCategories",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseOfListOfCategory"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Category"
        ],
        "operationId": "Category_AddCategory",
        "requestBody": {
          "x-name": "category",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseOfListOfCategory"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Category"
        ],
        "operationId": "Category_UpdateCategory",
        "requestBody": {
          "x-name": "category",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseOfListOfCategory"
                }
              }
            }
          }
        }
      }
    },
    "/api/Category/admin/{id}": {
      "delete": {
        "tags": [
          "Category"
        ],
        "operationId": "Category_DeleteCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseOfListOfCategory"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Product/{id}": {
      "delete": {
        "tags": [
          "Product"
        ],
        "operationId": "Product_DeleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseOfBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Product/{productId}": {
      "get": {
        "tags": [
          "Product"
        ],
        "operationId": "Product_GetProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseOfProduct"
                }
              }
            }
          }
        }
      }
    },
    "/api/Product/category/{categoryUrl}": {
      "get": {
        "tags": [
          "Product"
        ],
        "operationId": "Product_GetProductsByCategory",
        "parameters": [
          {
            "name": "categoryUrl",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseOfListOfProduct"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ProductType": {
      "get": {
        "tags": [
          "ProductType"
        ],
        "operationId": "ProductType_GetProductTypes",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseOfListOfProductType"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProductType"
        ],
        "operationId": "ProductType_AddProductType",
        "requestBody": {
          "x-name": "productType",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseOfListOfProductType"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProductType"
        ],
        "operationId": "ProductType_UpdateProductType",
        "requestBody": {
          "x-name": "productType",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductType"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseOfListOfProductType"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ServiceResponseOfInteger": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "integer",
            "format": "int32"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "UserRegister": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "maxLength": 100,
            "minLength": 6
          },
          "confirmPassword": {
            "type": "string"
          }
        }
      },
      "ServiceResponseOfString": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "UserLogin": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "ServiceResponseOfBoolean": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "boolean"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ServiceResponseOfListOfCategory": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "nullable": true,
            "items": {
              "nullable": true,
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Category"
                }
              ]
            }
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Category": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "visible": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          },
          "editing": {
            "type": "boolean"
          },
          "isNew": {
            "type": "boolean"
          }
        }
      },
...
      "ServiceResponseOfListOfProduct": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "nullable": true,
            "items": {
              "nullable": true,
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Product"
                }
              ]
            }
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Product": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          },
          "category": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Category"
              }
            ]
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "featured": {
            "type": "boolean"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariant"
            }
          },
          "visible": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          },
          "editing": {
            "type": "boolean"
          },
          "isNew": {
            "type": "boolean"
          }
        }
      },
      "Image": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "string"
          }
        }
      },
      "ProductVariant": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "productType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ProductType"
              }
            ]
          },
          "productTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "decimal"
          },
          "originalPrice": {
            "type": "number",
            "format": "decimal"
          },
          "visible": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          },
          "editing": {
            "type": "boolean"
          },
          "isNew": {
            "type": "boolean"
          }
        }
      },
      "ProductType": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "editing": {
            "type": "boolean"
          },
          "isNew": {
            "type": "boolean"
          }
        }
      },
      "ServiceResponseOfProduct": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Product"
              }
            ]
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ServiceResponseOfListOfProductType": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "nullable": true,
            "items": {
              "nullable": true,
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ProductType"
                }
              ]
            }
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  }
}